+++ EmptyCopy.qml
@@ -1,18 +1,18 @@
 <!DOCTYPE qgis PUBLIC 'http://mrcc.com/qgis.dtd' 'SYSTEM'>
-<qgis styleCategories="AllStyleCategories" minScale="100000000" labelsEnabled="0" version="3.40.5-Bratislava" simplifyAlgorithm="0" simplifyLocal="1" simplifyDrawingHints="1" readOnly="0" hasScaleBasedVisibilityFlag="0" symbologyReferenceScale="-1" maxScale="0" autoRefreshMode="Disabled" simplifyDrawingTol="1" autoRefreshTime="0" simplifyMaxScale="1">
+<qgis simplifyAlgorithm="0" symbologyReferenceScale="-1" readOnly="0" autoRefreshTime="0" labelsEnabled="0" simplifyDrawingTol="1" styleCategories="AllStyleCategories" hasScaleBasedVisibilityFlag="0" simplifyDrawingHints="1" maxScale="0" minScale="100000000" autoRefreshMode="Disabled" simplifyMaxScale="1" version="3.40.5-Bratislava" simplifyLocal="1">
   <flags>
     <Identifiable>1</Identifiable>
     <Removable>1</Removable>
     <Searchable>1</Searchable>
     <Private>0</Private>
   </flags>
-  <temporal endField="" limitMode="0" mode="0" durationField="" startField="" enabled="0" fixedDuration="0" durationUnit="min" startExpression="" endExpression="" accumulate="0">
+  <temporal mode="0" limitMode="0" startField="" enabled="0" accumulate="0" endField="" durationUnit="min" fixedDuration="0" startExpression="" endExpression="" durationField="">
     <fixedRange>
       <start></start>
       <end></end>
     </fixedRange>
   </temporal>
-  <elevation extrusion="0" type="IndividualFeatures" zoffset="0" respectLayerSymbol="1" zscale="1" extrusionEnabled="0" clamping="Terrain" binding="Centroid" symbology="Line" showMarkerSymbolInSurfacePlots="0">
+  <elevation zscale="1" symbology="Line" extrusion="0" clamping="Terrain" binding="Centroid" type="IndividualFeatures" zoffset="0" extrusionEnabled="0" showMarkerSymbolInSurfacePlots="0" respectLayerSymbol="1">
     <data-defined-properties>
       <Option type="Map">
         <Option type="QString" value="" name="name"/>
@@ -21,7 +21,7 @@
       </Option>
     </data-defined-properties>
     <profileLineSymbol>
-      <symbol alpha="1" clip_to_extent="1" type="line" frame_rate="10" force_rhr="0" is_animated="0" name="">
+      <symbol force_rhr="0" type="line" alpha="1" clip_to_extent="1" is_animated="0" name="" frame_rate="10">
         <data_defined_properties>
           <Option type="Map">
             <Option type="QString" value="" name="name"/>
@@ -29,7 +29,7 @@
             <Option type="QString" value="collection" name="type"/>
           </Option>
         </data_defined_properties>
-        <layer class="SimpleLine" locked="0" pass="0" enabled="1" id="{313dc7bd-a673-41fd-866e-908da41309ae}">
+        <layer locked="0" class="SimpleLine" enabled="1" id="{47b21789-ae32-4993-b392-127fb38606b3}" pass="0">
           <Option type="Map">
             <Option type="QString" value="0" name="align_dash_pattern"/>
             <Option type="QString" value="square" name="capstyle"/>
@@ -41,7 +41,7 @@
             <Option type="QString" value="MM" name="dash_pattern_offset_unit"/>
             <Option type="QString" value="0" name="draw_inside_polygon"/>
             <Option type="QString" value="bevel" name="joinstyle"/>
-            <Option type="QString" value="41,144,144,255,hsv:0.5,0.71372549019607845,0.56470588235294117,1" name="line_color"/>
+            <Option type="QString" value="222,98,238,255,hsv:0.81388888888888888,0.58823529411764708,0.93333333333333335,1" name="line_color"/>
             <Option type="QString" value="solid" name="line_style"/>
             <Option type="QString" value="0.6" name="line_width"/>
             <Option type="QString" value="MM" name="line_width_unit"/>
@@ -70,7 +70,7 @@
       </symbol>
     </profileLineSymbol>
     <profileFillSymbol>
-      <symbol alpha="1" clip_to_extent="1" type="fill" frame_rate="10" force_rhr="0" is_animated="0" name="">
+      <symbol force_rhr="0" type="fill" alpha="1" clip_to_extent="1" is_animated="0" name="" frame_rate="10">
         <data_defined_properties>
           <Option type="Map">
             <Option type="QString" value="" name="name"/>
@@ -78,15 +78,15 @@
             <Option type="QString" value="collection" name="type"/>
           </Option>
         </data_defined_properties>
-        <layer class="SimpleFill" locked="0" pass="0" enabled="1" id="{f39b992f-37c6-4005-b6c6-3c590a16bcfa}">
+        <layer locked="0" class="SimpleFill" enabled="1" id="{c93554b5-7d20-4136-b525-3ee18bcbfd8b}" pass="0">
           <Option type="Map">
             <Option type="QString" value="3x:0,0,0,0,0,0" name="border_width_map_unit_scale"/>
-            <Option type="QString" value="41,144,144,255,hsv:0.5,0.71372549019607845,0.56470588235294117,1" name="color"/>
+            <Option type="QString" value="222,98,238,255,hsv:0.81388888888888888,0.58823529411764708,0.93333333333333335,1" name="color"/>
             <Option type="QString" value="bevel" name="joinstyle"/>
             <Option type="QString" value="0,0" name="offset"/>
             <Option type="QString" value="3x:0,0,0,0,0,0" name="offset_map_unit_scale"/>
             <Option type="QString" value="MM" name="offset_unit"/>
-            <Option type="QString" value="29,103,103,255,hsv:0.5,0.71372549019607845,0.40335698481727322,1" name="outline_color"/>
+            <Option type="QString" value="158,70,170,255,hsv:0.81388888888888888,0.58823529411764708,0.66666666666666663,1" name="outline_color"/>
             <Option type="QString" value="solid" name="outline_style"/>
             <Option type="QString" value="0.2" name="outline_width"/>
             <Option type="QString" value="MM" name="outline_width_unit"/>
@@ -103,7 +103,7 @@
       </symbol>
     </profileFillSymbol>
     <profileMarkerSymbol>
-      <symbol alpha="1" clip_to_extent="1" type="marker" frame_rate="10" force_rhr="0" is_animated="0" name="">
+      <symbol force_rhr="0" type="marker" alpha="1" clip_to_extent="1" is_animated="0" name="" frame_rate="10">
         <data_defined_properties>
           <Option type="Map">
             <Option type="QString" value="" name="name"/>
@@ -111,18 +111,18 @@
             <Option type="QString" value="collection" name="type"/>
           </Option>
         </data_defined_properties>
-        <layer class="SimpleMarker" locked="0" pass="0" enabled="1" id="{a68cffe2-1495-4caa-baac-0f02413ae483}">
+        <layer locked="0" class="SimpleMarker" enabled="1" id="{8961618f-e0df-40ca-8d85-699a710aad5f}" pass="0">
           <Option type="Map">
             <Option type="QString" value="0" name="angle"/>
             <Option type="QString" value="square" name="cap_style"/>
-            <Option type="QString" value="41,144,144,255,hsv:0.5,0.71372549019607845,0.56470588235294117,1" name="color"/>
+            <Option type="QString" value="222,98,238,255,hsv:0.81388888888888888,0.58823529411764708,0.93333333333333335,1" name="color"/>
             <Option type="QString" value="1" name="horizontal_anchor_point"/>
             <Option type="QString" value="bevel" name="joinstyle"/>
             <Option type="QString" value="diamond" name="name"/>
             <Option type="QString" value="0,0" name="offset"/>
             <Option type="QString" value="3x:0,0,0,0,0,0" name="offset_map_unit_scale"/>
             <Option type="QString" value="MM" name="offset_unit"/>
-            <Option type="QString" value="29,103,103,255,hsv:0.5,0.71372549019607845,0.40335698481727322,1" name="outline_color"/>
+            <Option type="QString" value="158,70,170,255,hsv:0.81388888888888888,0.58823529411764708,0.66666666666666663,1" name="outline_color"/>
             <Option type="QString" value="solid" name="outline_style"/>
             <Option type="QString" value="0.2" name="outline_width"/>
             <Option type="QString" value="3x:0,0,0,0,0,0" name="outline_width_map_unit_scale"/>
@@ -144,9 +144,9 @@
       </symbol>
     </profileMarkerSymbol>
   </elevation>
-  <renderer-v2 type="singleSymbol" symbollevels="0" referencescale="-1" forceraster="0" enableorderby="0">
+  <renderer-v2 forceraster="0" enableorderby="0" type="singleSymbol" symbollevels="0" referencescale="-1">
     <symbols>
-      <symbol alpha="1" clip_to_extent="1" type="marker" frame_rate="10" force_rhr="0" is_animated="0" name="0">
+      <symbol force_rhr="0" type="marker" alpha="1" clip_to_extent="1" is_animated="0" name="0" frame_rate="10">
         <data_defined_properties>
           <Option type="Map">
             <Option type="QString" value="" name="name"/>
@@ -154,11 +154,11 @@
             <Option type="QString" value="collection" name="type"/>
           </Option>
         </data_defined_properties>
-        <layer class="SimpleMarker" locked="0" pass="0" enabled="1" id="{d56d254e-0af4-44e9-8330-a6592e49850c}">
+        <layer locked="0" class="SimpleMarker" enabled="1" id="{f4a40639-25f7-4fbb-add9-02b47d08176d}" pass="0">
           <Option type="Map">
             <Option type="QString" value="0" name="angle"/>
             <Option type="QString" value="square" name="cap_style"/>
-            <Option type="QString" value="184,53,158,255,hsv:0.8666666666666667,0.71372549019607845,0.72156862745098038,1" name="color"/>
+            <Option type="QString" value="146,45,99,255,hsv:0.91111111111111109,0.69411764705882351,0.5725490196078431,1" name="color"/>
             <Option type="QString" value="1" name="horizontal_anchor_point"/>
             <Option type="QString" value="bevel" name="joinstyle"/>
             <Option type="QString" value="circle" name="name"/>
@@ -211,7 +211,7 @@
     </activeChecks>
     <checkConfiguration/>
   </geometryOptions>
-  <legend showLabelLegend="0" type="default-vector"/>
+  <legend type="default-vector" showLabelLegend="0"/>
   <referencedLayers/>
   <fieldConfiguration>
     <field configurationFlags="NoFlag" name="name">
@@ -234,28 +234,28 @@
     <alias field="value" index="1" name=""/>
   </aliases>
   <splitPolicies>
-    <policy policy="Duplicate" field="name"/>
-    <policy policy="Duplicate" field="value"/>
+    <policy field="name" policy="Duplicate"/>
+    <policy field="value" policy="Duplicate"/>
   </splitPolicies>
   <duplicatePolicies>
-    <policy policy="Duplicate" field="name"/>
-    <policy policy="Duplicate" field="value"/>
+    <policy field="name" policy="Duplicate"/>
+    <policy field="value" policy="Duplicate"/>
   </duplicatePolicies>
   <defaults>
-    <default applyOnUpdate="0" field="name" expression="'Default Name'"/>
-    <default applyOnUpdate="0" field="value" expression=""/>
+    <default field="name" applyOnUpdate="0" expression="'Default Name'"/>
+    <default field="value" applyOnUpdate="0" expression=""/>
   </defaults>
   <constraints>
-    <constraint notnull_strength="0" unique_strength="0" exp_strength="0" constraints="0" field="name"/>
-    <constraint notnull_strength="0" unique_strength="0" exp_strength="0" constraints="0" field="value"/>
+    <constraint notnull_strength="0" exp_strength="0" unique_strength="0" field="name" constraints="0"/>
+    <constraint notnull_strength="0" exp_strength="0" unique_strength="0" field="value" constraints="0"/>
   </constraints>
   <constraintExpressions>
-    <constraint field="name" desc="" exp=""/>
-    <constraint field="value" desc="" exp=""/>
+    <constraint field="name" exp="" desc=""/>
+    <constraint field="value" exp="" desc=""/>
   </constraintExpressions>
   <expressionfields/>
   <attributeactions/>
-  <attributetableconfig sortOrder="0" sortExpression="" actionWidgetStyle="dropDown">
+  <attributetableconfig actionWidgetStyle="dropDown" sortOrder="0" sortExpression="">
     <columns/>
   </attributetableconfig>
   <conditionalstyles>
@@ -272,16 +272,16 @@
   <editorlayout>tablayout</editorlayout>
   <attributeEditorForm>
     <labelStyle overrideLabelFont="0" overrideLabelColor="0" labelColor="">
-      <labelFont description=".AppleSystemUIFont,13,-1,5,50,0,0,0,0,0" bold="0" strikethrough="0" style="" underline="0" italic="0"/>
+      <labelFont description=".AppleSystemUIFont,13,-1,5,50,0,0,0,0,0" style="" italic="0" strikethrough="0" bold="0" underline="0"/>
     </labelStyle>
-    <attributeEditorField showLabel="1" verticalStretch="0" horizontalStretch="0" index="0" name="name">
+    <attributeEditorField verticalStretch="0" showLabel="1" index="0" horizontalStretch="0" name="name">
       <labelStyle overrideLabelFont="0" overrideLabelColor="0" labelColor="">
-        <labelFont description=".AppleSystemUIFont,13,-1,5,50,0,0,0,0,0" bold="0" strikethrough="0" style="" underline="0" italic="0"/>
+        <labelFont description=".AppleSystemUIFont,13,-1,5,50,0,0,0,0,0" style="" italic="0" strikethrough="0" bold="0" underline="0"/>
       </labelStyle>
     </attributeEditorField>
-    <attributeEditorField showLabel="1" verticalStretch="0" horizontalStretch="0" index="1" name="value">
+    <attributeEditorField verticalStretch="0" showLabel="1" index="1" horizontalStretch="0" name="value">
       <labelStyle overrideLabelFont="0" overrideLabelColor="0" labelColor="">
-        <labelFont description=".AppleSystemUIFont,13,-1,5,50,0,0,0,0,0" bold="0" strikethrough="0" style="" underline="0" italic="0"/>
+        <labelFont description=".AppleSystemUIFont,13,-1,5,50,0,0,0,0,0" style="" italic="0" strikethrough="0" bold="0" underline="0"/>
       </labelStyle>
     </attributeEditorField>
   </attributeEditorForm>